refactor: ManagerMain 페이지에서 ReservationList 컴포넌트 분리
