+++ frontend/src/pages/ManagerMain/ManagerMain.tsx
@@ -4,31 +4,25 @@ import { useMutation } from 'react-query';
 import { useHistory, useLocation } from 'react-router-dom';
 import { deleteMap } from 'api/managerMap';
 import { deleteManagerReservation } from 'api/managerReservation';
-import { ReactComponent as DeleteIcon } from 'assets/svg/delete.svg';
-import { ReactComponent as EditIcon } from 'assets/svg/edit.svg';
 import { ReactComponent as MapEditorIcon } from 'assets/svg/map-editor.svg';
 import { ReactComponent as MenuIcon } from 'assets/svg/menu.svg';
 import { ReactComponent as SpaceEditorIcon } from 'assets/svg/space-editor.svg';
-import Button from 'components/Button/Button';
 import DateInput from 'components/DateInput/DateInput';
 import Header from 'components/Header/Header';
 import IconButton from 'components/IconButton/IconButton';
 import Layout from 'components/Layout/Layout';
 import PageHeader from 'components/PageHeader/PageHeader';
-import Panel from 'components/Panel/Panel';
-import ReservationListItem from 'components/ReservationListItem/ReservationListItem';
 import MESSAGE from 'constants/message';
 import PATH, { HREF } from 'constants/path';
-import useManagerMapReservations from 'hooks/query/useManagerMapReservations';
 import useManagerMaps from 'hooks/query/useManagerMaps';
 import useManagerSpaces from 'hooks/query/useManagerSpaces';
-import { Order, Reservation } from 'types/common';
+import { Reservation } from 'types/common';
 import { ErrorResponse, MapItemResponse } from 'types/response';
 import { formatDate } from 'utils/datetime';
-import { sortReservations } from 'utils/sort';
 import { isNullish } from 'utils/type';
 import * as Styled from './ManagerMain.styles';
 import MapDrawer from './units/MapDrawer';
+import ReservationList from './units/ReservationList';
 
 export interface ManagerMainState {
   mapId?: number;
@@ -47,7 +41,6 @@ const ManagerMain = (): JSX.Element => {
 
   const [selectedMapId, setSelectedMapId] = useState<number | null>(mapId ?? null);
   const [selectedMapName, setSelectedMapName] = useState('');
-  const [spacesOrder, setSpacesOrder] = useState<Order>(Order.Ascending);
 
   const onRequestError = (error: AxiosError<ErrorResponse>) => {
     alert(error.response?.data?.message ?? MESSAGE.MANAGER_MAIN.UNEXPECTED_GET_DATA_ERROR);
@@ -59,17 +52,6 @@ const ManagerMain = (): JSX.Element => {
 
   const organization = getMaps.data?.data.organization ?? '';
   const maps = useMemo((): MapItemResponse[] => getMaps.data?.data.maps ?? [], [getMaps]);
-
-  const getReservations = useManagerMapReservations(
-    {
-      mapId: selectedMapId as number,
-      date: formatDate(date),
-    },
-    {
-      enabled: !isNullish(selectedMapId),
-      onError: onRequestError,
-    }
-  );
 
   const getSpaces = useManagerSpaces(
     {
@@ -90,17 +72,7 @@ const ManagerMain = (): JSX.Element => {
     },
   });
 
-  const reservations = useMemo(() => getReservations.data?.data?.data ?? [], [getReservations]);
-  const sortedReservations = useMemo(
-    () => sortReservations(reservations, spacesOrder),
-    [reservations, spacesOrder]
-  );
-
   const spaces = useMemo(() => getSpaces.data?.data.spaces ?? [], [getSpaces]);
-
-  const handleClickSpacesOrder = () => {
-    setSpacesOrder((prev) => (prev === Order.Ascending ? Order.Descending : Order.Ascending));
-  };
 
   const removeMap = useMutation(deleteMap, {
     onSuccess: () => {
@@ -259,82 +231,13 @@ const ManagerMain = (): JSX.Element => {
           <DateInput date={date} setDate={setDate} />
         </Styled.DateInputWrapper>
 
-        {getReservations.isLoading && (
-          <Styled.NoticeWrapper>
-            <Styled.NoticeMessage>공간을 로딩 중입니다</Styled.NoticeMessage>
-          </Styled.NoticeWrapper>
-        )}
-
-        {!getReservations.isLoading &&
-          !reservations.length &&
-          (selectedMapId === null ? (
-            <Styled.NoticeWrapper>
-              <Styled.NoticeMessage>생성한 맵이 없습니다.</Styled.NoticeMessage>
-              <Styled.NoticeLink to={PATH.MANAGER_MAP_CREATE}>맵 생성하러 가기</Styled.NoticeLink>
-            </Styled.NoticeWrapper>
-          ) : (
-            <Styled.NoticeWrapper>
-              <Styled.NoticeMessage>생성한 공간이 없습니다.</Styled.NoticeMessage>
-              <Styled.NoticeLink to={HREF.MANAGER_SPACE_EDIT(selectedMapId)}>
-                공간 생성하러 가기
-              </Styled.NoticeLink>
-            </Styled.NoticeWrapper>
-          ))}
-        <Styled.ReservationsContainer>
-          <Styled.SpacesOrderButton variant="text" onClick={handleClickSpacesOrder}>
-            {spacesOrder === 'ascending' ? '오름차순 △' : '내림차순 ▽'}
-          </Styled.SpacesOrderButton>
-          <Styled.SpaceList>
-            {sortedReservations &&
-              sortedReservations.map(({ spaceId, spaceName, spaceColor, reservations }, index) => (
-                <Panel key={`space-${spaceId}`} role="listitem">
-                  <Panel.Header dotColor={spaceColor}>
-                    <Styled.PanelHeadWrapper>
-                      <Panel.Title>{spaceName}</Panel.Title>
-                      <Button
-                        variant="primary-text"
-                        size="dense"
-                        onClick={() => handleCreateReservation(spaceId)}
-                      >
-                        예약 추가하기
-                      </Button>
-                    </Styled.PanelHeadWrapper>
-                  </Panel.Header>
-                  <Panel.Content>
-                    {reservations.length === 0 ? (
-                      <Styled.PanelMessage>등록된 예약이 없습니다</Styled.PanelMessage>
-                    ) : (
-                      <>
-                        {reservations.map((reservation) => (
-                          <ReservationListItem
-                            key={`reservation-${reservation.id}`}
-                            reservation={reservation}
-                            control={
-                              <Styled.IconButtonWrapper>
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleEditReservation(reservation, spaceId)}
-                                >
-                                  <EditIcon width="100%" height="100%" />
-                                </IconButton>
-
-                                <IconButton
-                                  size="small"
-                                  onClick={() => handleDeleteReservation(reservation.id, spaceId)}
-                                >
-                                  <DeleteIcon width="100%" height="100%" />
-                                </IconButton>
-                              </Styled.IconButtonWrapper>
-                            }
-                          />
-                        ))}
-                      </>
-                    )}
-                  </Panel.Content>
-                </Panel>
-              ))}
-          </Styled.SpaceList>
-        </Styled.ReservationsContainer>
+        <ReservationList
+          selectedMapId={selectedMapId as number}
+          date={date}
+          onCreateReservation={handleCreateReservation}
+          onEditReservation={handleEditReservation}
+          onDeleteReservation={handleDeleteReservation}
+        />
       </Layout>
 
       {selectedMapId && maps && (
